rename: place 관련 컴포넌트 폴더 구조 변경
- components/party/place -> components/placeCreate
- create 폴더 삭제
