+++ src/components/placeCreate/info/DateTimeInput.tsx
@@ -19,7 +19,7 @@ const DateTimeInput = ({ value, setValueHandler }: InputProps) => {
         locale='ko'
         calendarType='US'
         value={new Date(value)}
-        onChange={(v: Date) => setValueHandler('visit_date', v)}
+        onChange={(v: Date) => setValueHandler('visitDate', v)}
       />
       <HStack>
         <NumberInput
@@ -29,7 +29,7 @@ const DateTimeInput = ({ value, setValueHandler }: InputProps) => {
           max={23}
           value={new Date(value).getHours()}
           onChange={(valueString) =>
-            setValueHandler('visit_date', new Date(value).setHours(Number(valueString)))
+            setValueHandler('visitDate', new Date(value).setHours(Number(valueString)))
           }>
           <NumberInputField />
           <NumberInputStepper>
@@ -45,7 +45,7 @@ const DateTimeInput = ({ value, setValueHandler }: InputProps) => {
           max={59}
           value={new Date(value).getMinutes()}
           onChange={(valueString) =>
-            setValueHandler('visit_date', new Date(value).setMinutes(Number(valueString)))
+            setValueHandler('visitDate', new Date(value).setMinutes(Number(valueString)))
           }>
           <NumberInputField />
           <NumberInputStepper>
